fix: fetch route count on delete dialog open for accurate cascade warning
- Move route count fetch from marker save to delete dialog open
- Add loading state to delete dialog while fetching
- Remove routeCount prop, add markerId prop to DeleteMarkerDialog
- Update CHANGELOG.md with cascade delete feature documentation
- Ensure fresh data shown in delete warning

Addresses Copilot review comments in PR #551
