+++ CHANGELOG.md
@@ -7,12 +7,15 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Add create-pull-request skill for standardized PR workflow
+- Add cascade delete for routes when marker is deleted
+- Add route count warning in marker delete dialog
 
 ### Changed
 - Improve marker panel design: display full marker names without truncation
 - Reorganize marker metadata (icons, badges, duration) into compact layout
 - Add multi-line support for long marker names
 - Improve visual hierarchy in marker list and available markers components
+- Fetch route count when delete dialog opens for accurate cascade warning
 
 ### Fixed
 - Use GitHub App for release workflow
@@ -23,6 +26,7 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add XSS protection with DOMPurify for marker notes rendering
 - Fix filter button positioning for indicator dot
 - Improve accessibility with aria-label for AI badge icon
+- Ensure data integrity when deleting markers with associated routes
 
 ## [v1.1.1] - 2026-03-25
